+++ src/client_sdl/sdl_player.h
@@ -23,7 +23,7 @@ public:
     sdl_player(SDL2pp::Renderer &renderer,
                         bool localTeam, bool mainPlayer);
     void update(int x, int y, int size_w, int size_h, double angle,
-                float dt, bool isMoving, bool isJumping, bool onTurbo,
+                unsigned int dt, bool isMoving, bool isJumping, bool onTurbo,
                 bool facingLeft);
     void render(SDL2pp::Renderer &renderer);
     ~sdl_player() = default;
